created category

+++ src/app/api/deleteSale/[id]/route.ts
@@ -1,33 +1,42 @@
 import Sale from "@/lib/models/SaleModel";
-import { NextResponse } from "next/server";
+import { verifyToken } from "@/lib/tokenmanage/verifyToken";
+import { NextRequest, NextResponse } from "next/server";
 
-interface Params {
-  params: { id?: string };
-}
 
-export async function DELETE(req: Request, { params }: Params) {
+export async function DELETE(
+  req: NextRequest,
+  { params }: { params: Promise<{ id: string }> }
+): Promise<NextResponse> {
+  const { id } = await params;
   try {
-    const { id } = params;
+    const { role } = await verifyToken(req);
 
+    if (role != "admin") {
+      return NextResponse.json(
+        { success: false, message: "You are not allow to delete the sale" },
+        { status: 400 }
+      );
+    }
     if (!id) {
-      return NextResponse.json({
-        success: false,
-        message: "ID is required",
-      }, { status: 400 });
+      return NextResponse.json(
+        { success: false, message: "ID is required" },
+        { status: 400 }
+      );
     }
 
-    const deletedSale = await Sale.findByIdAndDelete(id);
+    const sale = await Sale.findByIdAndDelete(id);
 
-    if (!deletedSale) {
-      return NextResponse.json({
-        success: false,
-        message: "Sale not found",
-      }, { status: 404 });
+    if (!sale) {
+      return NextResponse.json(
+        { success: false, message: "Sale not found" },
+        { status: 404 }
+      );
     }
 
     return NextResponse.json({
       success: true,
-      message: "Sale deleted successfully",
+      message: "sale deleted successfully",
+      saleDeletedCount: sale.deletedCount,
     });
   } catch (error: unknown) {
     let message = "Something went wrong";
